refactor: 폴더구조 변경 및 import 절대경로 선언#20
