+++ src/app/component/backtest/detail/Chart/Financial.tsx
