+++ src/app/component/backtest/detail/Chart/Pie.tsx
