+++ src/app/component/backtest/detail/StockSearch.tsx
